+++ tests/test_gamut.py
@@ -422,7 +422,15 @@ class TestRayTrace(util.ColorAsserts, unittest.TestCase):
         # When the intersect is not found, the last good intersect will be selected.
         self.assertColorEqual(
             Color('lch-d65', [0.1, 50, 270]).fit('srgb', method='raytrace', pspace='hct', adaptive=0.05),
-            Color('color(--lch-d65 0.08441 0.22909 34.545 / 1)')
+            Color('color(--lch-d65 0.1002 0.46923 14.39)')
+        )
+
+    def test_ray_trace_very_close_to_gamut(self):
+        """Test a scenario where a color is very close to the gamut."""
+
+        self.assertColorEqual(
+            Color('srgb', [1.0000000000000004, 0.19694944045886786, 0.026717808177518357]).fit(method='raytrace'),
+            Color('rgb(255 50.222 6.813)')
         )